backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
yfinance>=0.2.36
